Moved Search components to new directory structure
